Final Commit
I got the code working with the full dataset. The "Full_Dataset_CSV"
file is our dataset saved as a CSV (Big Surprise). Henry, I changed your
code back to how it was before. The number of rows in the database
corresponds to the number of pictures uploaded. Make sure you don't
change it back or it won't work. In order for the code to work when
uploading more than 20 pictures, you have to make a change to the
php.ini file for the wamp server. Search for the max_file_upload
varaible and make it higher, I set it to 1000. I'm not sure if you still
need to do this step when not using wamp as the server. Either way, it
should be specificied in the installation instructions. Henry is working
on a way to make that change in the code itself, but if he doesn't
figure it out, its a simple change that can be done easily. I'm done
doing any more work for this project. So like the title says, this is my
last commit.